Refactor to new folder
