+++ src/main/java/org/example/controller/BookController.java
@@ -1,11 +1,15 @@
-package org.example;
+package org.example.controller;
 
+import org.example.dto.Book;
+import org.example.repository.BookRepository;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
-@RestController("api")
+@RestController
+@RequestMapping("api")
 public class BookController {
 
     private final BookRepository bookRepository;
@@ -17,7 +21,7 @@ public class BookController {
     @GetMapping("books")
     public List<Book> getAllBooks() {
         return bookRepository.findAll().stream()
-                .map(book -> new Book(book.getTitle(), book.getAuthor(), book.getGenre(), book.getRating())).toList();
+                .map(book -> new Book(book.getTitle(), book.getGenre(), book.getRating(), book.getAuthor(), book.getLanguage())).toList();
     }
 
 }
