+++ src/main/java/org/example/dto/Author.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.dto;
 
 public record Author(String firstName, String lastName) {
 }
