+++ src/main/java/org/example/dto/Language.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.dto;
 
 public record Language(String textLanguage) {
 }
