+++ src/main/java/org/example/repository/AuthorRepository.java
@@ -1,7 +1,9 @@
-package org.example;
+package org.example.repository;
 
 import org.example.entities.Author;
 import org.springframework.data.repository.ListCrudRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface AuthorRepository extends ListCrudRepository<Author,Integer> {
 }
