+++ src/main/java/org/example/repository/BookRepository.java
@@ -1,10 +1,9 @@
-package org.example;
+package org.example.repository;
 
 import org.example.entities.Book;
 import org.springframework.data.repository.ListCrudRepository;
+import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
-
+@Repository
 public interface BookRepository extends ListCrudRepository<Book,Integer> {
-Optional<Book> findBookByAuthor(String author);
 }
