+++ src/main/java/org/example/repository/LanguageRepository.java
@@ -1,7 +1,9 @@
-package org.example;
+package org.example.repository;
 
 import org.example.entities.Language;
 import org.springframework.data.repository.ListCrudRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface LanguageRepository extends ListCrudRepository<Language,Integer> {
 }
